LV2: Completely redone events for instruments
The most visible change is that ev:EventPort is now deprecated.

The bigger change, however, is that LMMS now makes use of the new
note event extension.  This allows instruments that support it to
utilize certain features of LMMS, such as pitch automation and
panning for individual notes.  Note events are converted to MIDI
events for plugins that do not support the extension.